exception

+++ REST APIs/Product-Management-API/src/main/java/com/adminPanel/app/serviceLayer/ProductServiceImpl.java
@@ -1,6 +1,7 @@
 package com.adminPanel.app.serviceLayer;
 
 import com.adminPanel.app.dao.ProductDao;
+import com.adminPanel.app.exception.ProductNotFoundException;
 import com.adminPanel.app.model.Product;
 import com.adminPanel.app.model.ProductDetails;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -27,12 +28,28 @@ public class ProductServiceImpl implements ProductService
     }
 
 
-    public Product findById(int id) {
-        return productDAO.findByID(id);
+    public Product findById(int id) throws ProductNotFoundException {
+        if(id<=0 || id>1000)
+            throw new ProductNotFoundException("this id is invalid");
+        else {
+            Product product = productDAO.findByID(id);
+            if(product==null)
+                throw new ProductNotFoundException("there is no product in this id"+id);
+            else
+                return product;
+        }
     }
 
-    public void deleteById(int id) {
-         productDAO.deleteByID(id);
+    public void deleteById(int id)throws ProductNotFoundException {
+        if(id<=0 || id>1000)
+            throw new ProductNotFoundException("this id is invalid");
+        else {
+            Product product = productDAO.findByID(id);
+            if(product==null)
+                throw new ProductNotFoundException("there is no product in this id"+id);
+            else
+                productDAO.deleteByID(id);
+        }
     }
 
 
@@ -59,7 +76,15 @@ public class ProductServiceImpl implements ProductService
     }
 
     @Transactional
-    public List<Product> findByName(String searchKey) {
-        return productDAO.findByName(searchKey);
+    public List<Product> findByName(String searchKey) throws ProductNotFoundException {
+        if(searchKey=="")
+            throw new ProductNotFoundException("where is the product name!");
+        else{
+            List<Product> productList= productDAO.findByName(searchKey);
+            if(productList.isEmpty())
+                throw new ProductNotFoundException("there is no product in this name");
+            else
+                return productList;
+        }
     }
 }
